Move formatter config types into fmt module
Refs #446
