+++ crates/djls-conf/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod diagnostics;
+pub mod fmt;
 pub mod tagspecs;
 
 use std::fs;
-use std::num::NonZeroU16;
 use std::path::Path;
 
 use anyhow::Context;
@@ -19,6 +19,9 @@ use thiserror::Error;
 
 pub use crate::diagnostics::DiagnosticSeverity;
 pub use crate::diagnostics::DiagnosticsConfig;
+pub use crate::fmt::ContentType;
+pub use crate::fmt::FormatConfig;
+pub use crate::fmt::IndentStyle;
 pub use crate::tagspecs::ArgKindDef;
 pub use crate::tagspecs::ArgTypeDef;
 pub use crate::tagspecs::EndTagDef;
@@ -63,77 +66,6 @@ pub enum ConfigError {
     PyprojectParse(#[from] toml::de::Error),
     #[error("Failed to serialize extracted pyproject.toml data")]
     PyprojectSerialize(#[from] toml::ser::Error),
-}
-
-#[derive(Debug, Deserialize, Clone, Copy, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum IndentStyle {
-    #[default]
-    Spaces,
-    Tabs,
-}
-
-#[derive(Debug, Deserialize, Clone, Copy, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum ContentType {
-    #[default]
-    Auto,
-    Html,
-    Text,
-}
-
-#[derive(Debug, Deserialize, Clone, PartialEq, Eq)]
-#[serde(default)]
-pub struct FormatConfig {
-    #[serde(deserialize_with = "deserialize_nonzero_u16")]
-    indent_width: NonZeroU16,
-    indent_style: IndentStyle,
-    content_type: ContentType,
-    #[serde(deserialize_with = "deserialize_nonzero_u16")]
-    print_width: NonZeroU16,
-}
-
-impl Default for FormatConfig {
-    fn default() -> Self {
-        Self {
-            indent_width: NonZeroU16::new(4).expect("format default indent_width is non-zero"),
-            indent_style: IndentStyle::Spaces,
-            content_type: ContentType::Auto,
-            print_width: NonZeroU16::new(80).expect("format default print_width is non-zero"),
-        }
-    }
-}
-
-impl FormatConfig {
-    #[must_use]
-    pub fn indent_width(&self) -> u16 {
-        self.indent_width.get()
-    }
-
-    #[must_use]
-    pub fn indent_style(&self) -> IndentStyle {
-        self.indent_style
-    }
-
-    #[must_use]
-    pub fn content_type(&self) -> ContentType {
-        self.content_type
-    }
-
-    #[must_use]
-    pub fn print_width(&self) -> u16 {
-        self.print_width.get()
-    }
-}
-
-fn deserialize_nonzero_u16<'de, D>(deserializer: D) -> Result<NonZeroU16, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    use serde::de::Error;
-
-    let value = u16::deserialize(deserializer)?;
-    NonZeroU16::new(value).ok_or_else(|| D::Error::custom("expected a non-zero integer"))
 }
 
 #[derive(Debug, Deserialize, Default, PartialEq, Clone)]
@@ -489,12 +421,18 @@ content_type = "text"
             .unwrap();
 
             let settings = Settings::new(Utf8Path::from_path(dir.path()).unwrap(), None).unwrap();
+            assert_eq!(settings.format().content_type(), ContentType::Text);
             assert_eq!(
-                settings.format(),
-                &FormatConfig {
-                    content_type: ContentType::Text,
-                    ..FormatConfig::default()
-                }
+                settings.format().indent_width(),
+                FormatConfig::default().indent_width()
+            );
+            assert_eq!(
+                settings.format().indent_style(),
+                FormatConfig::default().indent_style()
+            );
+            assert_eq!(
+                settings.format().print_width(),
+                FormatConfig::default().print_width()
             );
         }
